requirements.txt
openai
pandas
numpy
scikit-learn
selenium
python-dotenv
webdriver-manager
matplotlib
seaborn